Create assignment for landing page
This is the first iteration of the assignment for CSS with The Odin
Project. The Assignment submission includes the index and styles files
as well as a simple readme. This has given me practice with website
design including the extensive use of flex. I've been able to google
other things that weren't covered in this course, such as button design
and rounding corners, but that type of thing is to be expected as there
is so much that can be done in CSS and there is such a limited time in
this project.

Follow-on work for this includes developing better practice for writing
the style sheet to reduce reducndancies in styling and better design of
cascading divs and classes to minimize the need for extraneous classes
and containers.

To continue practice for this assignment, I will begin to work on my
own website which will be to track my 50-states marathon goal.